move common code to the layer and create setup.py so that it can be pip installed into the layer venv
