--- api/terraform/python/lambda_langchain/requirements.txt
@@ -1,14 +1,0 @@
-# -----------------------------------------------------------------------------
-# written by: Lawrence McDaniel
-#             https://lawrencemcdaniel.com
-#
-# usage:      Python requirements for AWS Lambda functions. Create a virtual
-#             environment in the root of this repository named `.venv`. Terraform
-#             modules will look for and include these requirements in the zip
-#             packages for each Python-based Lambda function.
-# -----------------------------------------------------------------------------
-python-dotenv
-openai>=0.28.0
-langchain>=0.0.323
-langchain-experimental>=0.0.34
-pinecone-client>=2.2.4
