+++ api/terraform/python/lambda_langchain/tests/events/test_01.request.json
