+++ api/terraform/python/layer_genai/.gitignore
@@ -1,3 +1,4 @@
 python
-layer_langchain_dst.zip
+layer_genai_dst.zip
 venv
+genai
