+++ api/terraform/python/layer_genai/openai_utils/README.md
@@ -1,0 +1,3 @@
+# OpenAI Utils
+
+Common constants, validations and misc utility functions that are call by both the openai and langchain lambdas.
