+++ api/terraform/python/layer_genai/openai_utils/data/example_request.json
@@ -1,6 +1,8 @@
 {
   "model": "gpt-3.5-turbo",
   "end_point": "ChatCompletion",
+  "temperature": 0.7,
+  "max_tokens": 100,
   "messages": [
       {
           "role": "system",
