+++ api/terraform/python/layer_genai/openai_utils/openai_utils.egg-info/PKG-INFO
@@ -1,0 +1,6 @@
+Metadata-Version: 2.1
+Name: openai-utils
+Version: 0.1.0
+Summary: Common utilities for OpenAI
+Author: Lawrence McDaniel
+Author-email: lpm0073@gmail.com
