+++ api/terraform/python/layer_genai/openai_utils/openai_utils.egg-info/SOURCES.txt
@@ -1,0 +1,6 @@
+README.md
+setup.py
+openai_utils.egg-info/PKG-INFO
+openai_utils.egg-info/SOURCES.txt
+openai_utils.egg-info/dependency_links.txt
+openai_utils.egg-info/top_level.txt
