+++ api/terraform/python/layer_genai/openai_utils/setup.py
@@ -1,0 +1,11 @@
+from setuptools import setup, find_packages
+
+setup(
+    name="openai_utils",
+    version="0.1.0",
+    description="Common utilities for OpenAI",
+    author="Lawrence McDaniel",
+    author_email="lpm0073@gmail.com",
+    packages=find_packages(),
+    install_requires=[],
+)
