+++ api/terraform/python/layer_genai/requirements.txt
@@ -12,3 +12,6 @@ openai>=0.28.0
 langchain>=0.0.323
 langchain-experimental>=0.0.34
 pinecone-client>=2.2.4
+
+# local packages
+./openai_utils
