+++ api/terraform/scripts/create_pkg_lambda_langchain.sh
@@ -36,4 +36,3 @@ mkdir -p $PACKAGE_FOLDER
 
 # copy the python module(s) to the package folder
 cp lambda_handler.py $PACKAGE_FOLDER/
-cp -R $PACKAGE_NAME $PACKAGE_FOLDER/
